extracting PredicateCallGraph object for reuse
